+++ src/Event/ForkEvent.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace TK\GitHubWebhook\Event;
+
+use TK\GitHubWebhook\Model\InstallationLite;
+use TK\GitHubWebhook\Model\Repository;
+use TK\GitHubWebhook\Model\User;
+use TK\GitHubWebhook\Util;
+
+class ForkEvent extends AbstractEvent
+{
+    public Repository $forkee;
+    public InstallationLite|null $installation;
+
+    public static function fromArray(array $data): ForkEvent
+    {
+        $repository = Util::getArgSafe($data, "repository", Repository::fromArray(...));
+        $sender = Util::getArgSafe($data, "sender", User::fromArray(...));
+        $organization = Util::getArgSafe($data, "organization", User::fromArray(...));
+
+        $instance = new ForkEvent($repository, $sender, $organization);
+        $instance->forkee = Util::getArgSafe($data, "forkee", Repository::fromArray(...));
+        $instance->installation = Util::getArgSafe($data, "installation", InstallationLite::fromArray(...));
+        return $instance;
+    }
+}