+++ src/Event/PushEvent.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace TK\GitHubWebhook\Event;
+
+use TK\GitHubWebhook\Model\Commit;
+use TK\GitHubWebhook\Model\Committer;
+use TK\GitHubWebhook\Model\InstallationLite;
+use TK\GitHubWebhook\Model\Repository;
+use TK\GitHubWebhook\Model\User;
+use TK\GitHubWebhook\Util;
+
+class PushEvent extends AbstractEvent
+{
+    public string $ref;
+    public string $before;
+    public string $after;
+    public bool $created;
+    public bool $deleted;
+    public bool $forced;
+    public string|null $base_ref;
+    public string $compare;
+    /** @var Commit[] $commits */
+    public array $commits;
+    public Commit|null $head_commit;
+    public Committer $pusher;
+    public InstallationLite|null $installation;
+
+    public static function fromArray(array $data): PushEvent
+    {
+        $repository = Util::getArgSafe($data, "repository", Repository::fromArray(...));
+        $sender = Util::getArgSafe($data, "sender", User::fromArray(...));
+        $organization = Util::getArgSafe($data, "organization", User::fromArray(...));
+
+        $instance = new PushEvent($repository, $sender, $organization);
+        $instance->ref = $data["ref"];
+        $instance->before = $data["before"];
+        $instance->after = $data["after"];
+        $instance->created = $data["created"];
+        $instance->deleted = $data["deleted"];
+        $instance->forced = $data["forced"];
+        $instance->base_ref = $data["base_ref"] ?? null;
+        $instance->compare = $data["compare"];
+        $instance->commits = Util::getArraySafe($data, "commits", Commit::fromArray(...));
+        $instance->head_commit = Util::getArgSafe($data, "head_commit", Commit::fromArray(...));
+        $instance->pusher = Util::getArgSafe($data, "pusher", Committer::fromArray(...));
+        $instance->installation = Util::getArgSafe($data, "installation", InstallationLite::fromArray(...));
+        return $instance;
+    }
+}