+++ src/Model/Star/EventTypes.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace TK\GitHubWebhook\Model\Star;
+
+enum EventTypes: string
+{
+    case CREATED = "created";
+    case DELETED = "deleted";
+}