+++ src/Handler/EventNotFoundHandler.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TK\GitHubWebhook\Handler;
+
+use TK\GitHubWebhook\Response;
+
+interface EventNotFoundHandler
+{
+  public function handleEvent(string $event_type, array $json, Response $response): Response;
+}