+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - support for all `pull_request` events
 - support for the `fork` event
 - support for the `push` event
+- support for the `star` event
 
 ## 1.0.0 - 2024-07-25